vault-python/app/requirements.txt
hvac
psycopg2-binary
